Break down ui.ts into focused UI modules
Refactor the 806-line ui.ts file into smaller, focused modules for
improved navigability and maintainability:

- ui/helpers.ts: Utility functions (formatDuration, escapeHtml, etc.)
- ui/header.ts: renderHeader() for workspace header
- ui/loading.ts: renderLoadingState() for factory reset animation
- ui/terminal-primary.ts: renderPrimaryTerminal() and error states
- ui/terminal-grid.ts: renderMiniTerminals() and mini terminal cards
- ui/index.ts: Re-exports for backward compatibility

Existing imports from "./lib/ui" continue to work unchanged.

Closes #749
